+++ lib/Hakyll/Web/Feed.hs
@@ -133,12 +133,7 @@ renderFeed feedType feedTpl itemTpl config itemContext items = do
           JsonFeed -> ", "
 
     body <- makeItem =<< applyJoinTemplateList itemDelim itemTpl itemContext' protectedItems
-    rendered <- applyTemplate feedTpl feedContext body
-    -- when (feedType == XmlFeed) $ case parseText def $ T.pack (itemBody rendered) of
-    --   Right _ -> pure ()
-    --   Left err -> compilerThrow ["Generated feed contains invalid XML (perhaps you id not escape a metadata field?)",
-    --                              displayException err]
-    pure rendered
+    applyTemplate feedTpl feedContext body
   where
     applyFilter :: (Monad m,Functor f) => (String -> String) -> f String -> m (f String)
     applyFilter tr str = return $ fmap tr str
@@ -190,8 +185,14 @@ validateXMLFeed rendered = case parseText def $ T.pack (itemBody rendered) of
         ["Generated feed contains invalid XML (perhaps you id not escape a metadata field?)",
           displayException err]
 
+
 --------------------------------------------------------------------------------
 -- | Render an RSS feed using given templates with a number of items.
+--
+-- The resulting feed will not be validated. Prefer to use 'renderRssWithTemplates'
+-- when possible.
+--
+-- @since 4.16.7.0
 renderRssWithTemplatesNoValidate ::
        Template                -- ^ Feed template
     -> Template                -- ^ Item template
@@ -203,8 +204,11 @@ renderRssWithTemplatesNoValidate feedTemplate itemTemplate config context = rend
     XmlFeed feedTemplate itemTemplate config
     (makeItemContext "%a, %d %b %Y %H:%M:%S UT" context)
 
+
 --------------------------------------------------------------------------------
 -- | Render an RSS feed using given templates with a number of items.
+--
+-- The resulting RSS feed will be validated automatically.
 renderRssWithTemplates ::
        Template                -- ^ Feed template
     -> Template                -- ^ Item template
@@ -216,8 +220,14 @@ renderRssWithTemplates feedTemplate itemTemplate config context items =
   renderRssWithTemplatesNoValidate feedTemplate itemTemplate config context items
   >>= validateXMLFeed
 
+
 --------------------------------------------------------------------------------
 -- | Render an Atom feed using given templates with a number of items.
+--
+-- The resulting feed will not be validated. Prefer to use 'renderAtomWithTemplates'
+-- when possible.
+--
+-- @since 4.16.7.0
 renderAtomWithTemplatesNoValidate ::
        Template                -- ^ Feed template
     -> Template                -- ^ Item template
@@ -230,8 +240,11 @@ renderAtomWithTemplatesNoValidate feedTemplate itemTemplate config context items
     (makeItemContext "%Y-%m-%dT%H:%M:%SZ" context)
     items
 
+
 --------------------------------------------------------------------------------
 -- | Render an Atom feed using given templates with a number of items.
+--
+-- The resulting Atom feed will be validated automatically.
 renderAtomWithTemplates ::
        Template                -- ^ Feed template
     -> Template                -- ^ Item template
@@ -242,6 +255,7 @@ renderAtomWithTemplates ::
 renderAtomWithTemplates feedTemplate itemTemplate config context items =
   renderAtomWithTemplatesNoValidate feedTemplate itemTemplate config context items
   >>= validateXMLFeed
+
 
 --------------------------------------------------------------------------------
 -- | Render a JSON feed using given templates with a number of items.
@@ -259,14 +273,22 @@ renderJsonWithTemplates feedTemplate itemTemplate config context = renderFeed
 
 --------------------------------------------------------------------------------
 -- | Render an RSS feed with a number of items.
+--
+-- The resulting feed will not be validated. Prefer to use 'renderRss'
+-- when possible.
+--
+-- @since 4.16.7.0
 renderRssNoValidate :: FeedConfiguration       -- ^ Feed configuration
           -> Context String          -- ^ Item context
           -> [Item String]           -- ^ Feed items
           -> Compiler (Item String)  -- ^ Resulting feed
 renderRssNoValidate = renderRssWithTemplatesNoValidate rssTemplate rssItemTemplate
 
+
 --------------------------------------------------------------------------------
 -- | Render an RSS feed with a number of items.
+--
+-- The resulting RSS feed will be validated automatically.
 renderRss :: FeedConfiguration       -- ^ Feed configuration
           -> Context String          -- ^ Item context
           -> [Item String]           -- ^ Feed items
@@ -274,16 +296,25 @@ renderRss :: FeedConfiguration       -- ^ Feed configuration
 renderRss config context items = renderRssNoValidate config context items
   >>= validateXMLFeed
 
+
 --------------------------------------------------------------------------------
 -- | Render an Atom feed with a number of items.
+--
+-- The resulting feed will not be validated. Prefer to use 'renderAtom'
+-- when possible.
+--
+-- @since 4.16.7.0
 renderAtomNoValidate :: FeedConfiguration       -- ^ Feed configuration
            -> Context String          -- ^ Item context
            -> [Item String]           -- ^ Feed items
            -> Compiler (Item String)  -- ^ Resulting feed
 renderAtomNoValidate = renderAtomWithTemplatesNoValidate atomTemplate atomItemTemplate
 
+
 --------------------------------------------------------------------------------
 -- | Render an Atom feed with a number of items.
+--
+-- The resulting Atom feed will be validated automatically.
 renderAtom :: FeedConfiguration       -- ^ Feed configuration
            -> Context String          -- ^ Item context
            -> [Item String]           -- ^ Feed items